requirements.txt

streamlit
openai
openai-whisper
torch
ffmpeg-python
gtts
soundfile
av
streamlit-webrtc
